pyproject.toml
[project]
name = "stroke-classification"
version = "0.1.0"
description = "Stroke classification with deep learning and knowledge distillation"
readme = "README.md"
license = {text = "Apache-2.0"}
authors = [
    {name = "Melis Kılıç"},
    {name = "Esra Koç"},
]
requires-python = ">=3.13"
dependencies = [
    "hydra-core>=1.3.2",
    "matplotlib>=3.10.8",
    "numpy>=2.4.4",
    "omegaconf>=2.3.0",
    "opencv-python>=4.13.0.92",
    "pandas>=3.0.2",
    "pillow>=12.2.0",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "timm>=1.0.26",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "tqdm>=4.67.3",
    "wandb>=0.25.1",
]
